State Updates
Ironed out issues with the state machine fixed the bug that forced the AI into the idle state and added new functionality to the Run state. Still trying to put the Room change code into the states which should be easier now that the wander state can last longer than half a frame.